Notificaciones

+++ app/Http/Controllers/BrandController.php
@@ -9,7 +9,7 @@ use App\Models\Brand;
 class BrandController extends Controller
 {
 
-   /**
+    /**
      *La función de índice es una función que devuelve una vista de la página de índice de marcas
      *
      *@return Una vista llamada marcas.index con una variable llamada marcas
@@ -17,7 +17,7 @@ class BrandController extends Controller
     public function index()
     {
         $brands = Brand::paginate(5);
-        return view('brands.index', compact('brands'));
+        return view('brands.index', compact('brands'))->with('brand-success', 'La marca se ha guardado correctamente');
     }
 
     /**
@@ -37,20 +37,20 @@ class BrandController extends Controller
         return redirect()->route('brands.index')->with('brand-success', 'Marca creada correctamente');
     }
 
-   /**
-    *La función de edición toma la identificación de una marca y devuelve una vista con la marca
-    *
-    *@param id El id de la marca que queremos editar.
-    *
-    *@return Se devuelve la vista de edición.
-    */
+    /**
+     *La función de edición toma la identificación de una marca y devuelve una vista con la marca
+     *
+     *@param id El id de la marca que queremos editar.
+     *
+     *@return Se devuelve la vista de edición.
+     */
     public function edit($id)
     {
         $brand = Brand::find($id);
         return view('brands.edit', compact('brand'));
     }
 
-   /**
+    /**
      *Toma una solicitud, encuentra la marca con la identificación de la solicitud, actualiza la marca con el
      *nombre y referencia de la solicitud, y luego redirige a la página de índice
      *
@@ -64,11 +64,11 @@ class BrandController extends Controller
         $brand = Brand::find($id);
         $brand->name = $request->name;
         $brand->reference = $request->reference;
-        $brand->save();
+        $brand->update();
         return redirect()->route('brands.index');
     }
 
-   /**
+    /**
      *La función toma una identificación como parámetro, encuentra la marca con esa identificación, la elimina y luego
      *redirige a la página de índice
      *
@@ -79,7 +79,12 @@ class BrandController extends Controller
     public function destroy($id)
     {
         $brand = Brand::find($id);
-        $brand->delete();
-        return redirect()->route('brands.index');
+
+        if ($brand->products->count() > 0) {
+            return redirect()->route('brands.index')->with('brand-error', 'La marca no se puede eliminar porque tiene productos asociados');
+        } else {
+            $brand->delete();
+            return redirect()->route('brands.index')->with('brand-deleted', 'La marca se ha eliminado correctamente');
+        }
     }
 }
